backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
pydantic
pydantic-settings
google-auth
requests
PyJWT
supabase
pypdf
redis
httpx
email-validator
reportlab
bcrypt
google-generativeai
prometheus_client

# Test dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0